+++ backend/moonstream/actions.py
@@ -12,6 +12,13 @@ from sqlalchemy.orm import Session, query, query_expression
 
 from . import data
 from .settings import ETHERSCAN_SMARTCONTRACTS_BUCKET
+import uuid
+from bugout.data import BugoutResource
+from .settings import (
+    MOONSTREAM_APPLICATION_ID,
+    bugout_client as bc,
+    BUGOUT_REQUEST_TIMEOUT_SECONDS,
+)
 
 logger = logging.getLogger(__name__)
 ETHERSCAN_SMARTCONTRACT_LABEL_NAME = "etherscan_smartcontract"
@@ -141,3 +148,25 @@ def get_address_labels(
         )
 
     return addresses_response
+
+
+def create_onboarding_resource(
+    token: uuid.UUID,
+    resource_data: Dict[str, Any] = {
+        "type": data.USER_ONBOARDING_STATE,
+        "steps": {
+            "welcome": 0,
+            "subscriptions": 0,
+            "stream": 0,
+        },
+        "is_complete": False,
+    },
+) -> BugoutResource:
+
+    resource = bc.create_resource(
+        token=token,
+        application_id=MOONSTREAM_APPLICATION_ID,
+        resource_data=resource_data,
+        timeout=BUGOUT_REQUEST_TIMEOUT_SECONDS,
+    )
+    return resource
